Cache Consumer attributes field
