+++ pulsar-broker/src/main/java/org/apache/pulsar/broker/stats/OpenTelemetryConsumerStats.java
@@ -27,7 +27,6 @@ import org.apache.pulsar.broker.PulsarService;
 import org.apache.pulsar.broker.service.Consumer;
 import org.apache.pulsar.broker.service.Subscription;
 import org.apache.pulsar.broker.service.Topic;
-import org.apache.pulsar.common.naming.TopicName;
 import org.apache.pulsar.opentelemetry.OpenTelemetryAttributes;
 
 public class OpenTelemetryConsumerStats implements AutoCloseable {
@@ -123,39 +122,7 @@ public class OpenTelemetryConsumerStats implements AutoCloseable {
     }
 
     private void recordMetricsForConsumer(Consumer consumer) {
-        var subscription = consumer.getSubscription();
-        var topicName = TopicName.get(subscription.getTopic().getName());
-
-        var builder = Attributes.builder()
-                .put(OpenTelemetryAttributes.PULSAR_CONSUMER_NAME, consumer.consumerName())
-                .put(OpenTelemetryAttributes.PULSAR_CONSUMER_ID, consumer.consumerId())
-                .put(OpenTelemetryAttributes.PULSAR_CONSUMER_CONNECTED_SINCE,
-                        consumer.getConnectedSince().getEpochSecond())
-                .put(OpenTelemetryAttributes.PULSAR_SUBSCRIPTION_NAME, subscription.getName())
-                .put(OpenTelemetryAttributes.PULSAR_SUBSCRIPTION_TYPE, consumer.subType().toString())
-                .put(OpenTelemetryAttributes.PULSAR_DOMAIN, topicName.getDomain().toString())
-                .put(OpenTelemetryAttributes.PULSAR_TENANT, topicName.getTenant())
-                .put(OpenTelemetryAttributes.PULSAR_NAMESPACE, topicName.getNamespace())
-                .put(OpenTelemetryAttributes.PULSAR_TOPIC, topicName.getPartitionedTopicName());
-        if (topicName.isPartitioned()) {
-            builder.put(OpenTelemetryAttributes.PULSAR_PARTITION_INDEX, topicName.getPartitionIndex());
-        }
-        var clientAddress = consumer.getClientAddressAndPort();
-        if (clientAddress != null) {
-            builder.put(OpenTelemetryAttributes.PULSAR_CLIENT_ADDRESS, clientAddress);
-        }
-        var clientVersion = consumer.getClientVersion();
-        if (clientVersion != null) {
-            builder.put(OpenTelemetryAttributes.PULSAR_CLIENT_VERSION, clientVersion);
-        }
-        var metadataList = consumer.getMetadata()
-                .entrySet()
-                .stream()
-                .map(e -> String.format("%s:%s", e.getKey(), e.getValue()))
-                .toList();
-        builder.put(OpenTelemetryAttributes.PULSAR_CONSUMER_METADATA, metadataList);
-        var attributes = builder.build();
-
+        var attributes = consumer.getOpenTelemetryAttributes();
         messageOutCounter.record(consumer.getMsgOutCounter(), attributes);
         bytesOutCounter.record(consumer.getBytesOutCounter(), attributes);
         messageAckCounter.record(consumer.getMessageAckCounter(), attributes);
